+++ android/rctmgl/src/main/java-v10/com/mapbox/rctmgl/components/styles/sources/RCTMGLPointSource.kt
@@ -94,7 +94,7 @@ class RCTMGLPointSource(context: Context, private val mManager: RCTMGLPointSourc
         val fps = 30.0
         var ratio = 0.0
 
-        val durationSec = _mAnimationDuration / 1000.0
+        val durationSec = _mAnimationDuration.toDouble() / 1000.0
         val ratioIncr = 1.0 / (fps * durationSec)
         val period = 1000.0 / fps
